+++ src/main/java/com/kosa/ShareTour/repository/PackageRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.time.LocalDate;
 import java.util.List;
 
-public interface PackageRepository extends JpaRepository<Package, Integer>{
+public interface PackageRepository extends JpaRepository<Package, Long>{
 
 //    //패키지 제목으로 찾기(여럿)
 //    List<Package> findByPackageTitle(String title);
